+++ packages/FormAuto/FormAuto.vue
@@ -283,9 +283,9 @@ export default class ElFormAuto extends Vue {
 					let [sd, ed] = this.model[name] || [null, null];
 					data[sn] = sd;
 					data[en] = ed;
-					if (item.type == "daterange" && item.suffixTime == true) {
-						data[sn] += "00:00:00";
-						data[en] += "23:59:59";
+					if (sd && ed && item.type == "daterange" && item.suffixTime == true) {
+						data[sn] += " 00:00:00";
+						data[en] += " 23:59:59";
 					}
 				}
 			}
@@ -306,7 +306,6 @@ export default class ElFormAuto extends Vue {
 			let value = model[name];
 			let field = this.fields[name];
 			if (field && /radio|select|check/.test(field.type)) {
-				console.log(name, value)
 				if (field.type == "select" && field.remote) {
 					let values = this.selectEcho(name, value);
 					if (Array.isArray(value)) {
@@ -352,6 +351,8 @@ export default class ElFormAuto extends Vue {
 		} else if (options && options.label && options.value) {
 			field.echoOptions = [options];
 			return `${options.value}`;
+		} else {
+			return options
 		}
 	}